refactor(editor): Extract pin data logic into useWorkflowDocumentPinData composable
Move pin data state, actions, normalization functions, and event bus handling
from workflowDocument.store.ts into a dedicated composable for better
separation of concerns.
