+++ packages/frontend/editor-ui/src/app/stores/workflowDocument.store.ts
@@ -1,10 +1,10 @@
 import { defineStore, getActivePinia, type StoreGeneric } from 'pinia';
 import { STORES } from '@n8n/stores';
 import { ref, readonly, inject } from 'vue';
-import type { INodeExecutionData, IPinData } from 'n8n-workflow';
 import { WorkflowDocumentStoreKey } from '@/app/constants/injectionKeys';
-import { isJsonKeyObject, stringSizeInBytes } from '@/app/utils/typesUtils';
-import { dataPinningEventBus } from '@/app/event-bus';
+import { useWorkflowDocumentPinData } from './workflowDocument/useWorkflowDocumentPinData';
+
+export { getPinDataSize } from './workflowDocument/useWorkflowDocumentPinData';
 
 // Pinia internal type - _s is the store registry Map
 type PiniaInternal = ReturnType<typeof getActivePinia> & {
@@ -26,73 +26,13 @@ type SetTagsAction = Action<'setTags', { tags: string[] }>;
 type AddTagsAction = Action<'addTags', { tags: string[] }>;
 type RemoveTagAction = Action<'removeTag', { tagId: string }>;
 
-type SetPinDataAction = Action<'setPinData', { pinData: IPinData }>;
-type PinNodeDataAction = Action<'pinNodeData', { nodeName: string; data: INodeExecutionData[] }>;
-type UnpinNodeDataAction = Action<'unpinNodeData', { nodeName: string }>;
-type RenamePinDataNodeAction = Action<'renamePinDataNode', { oldName: string; newName: string }>;
-
-type WorkflowDocumentAction =
-	| SetTagsAction
-	| AddTagsAction
-	| RemoveTagAction
-	| SetPinDataAction
-	| PinNodeDataAction
-	| UnpinNodeDataAction
-	| RenamePinDataNodeAction;
+type TagAction = SetTagsAction | AddTagsAction | RemoveTagAction;
 
 /**
  * Gets the store ID for a workflow document store.
  */
 export function getWorkflowDocumentStoreId(id: string) {
 	return `${STORES.WORKFLOW_DOCUMENTS}/${id}`;
-}
-
-/**
- * Normalize an array of node execution data items by stripping runtime properties.
- * Only keeps json, binary, and pairedItem.
- */
-function normalizeNodeExecutionData(data: INodeExecutionData[]): INodeExecutionData[] {
-	if (!Array.isArray(data)) {
-		data = [data];
-	}
-
-	return data.map((item) => {
-		if (isJsonKeyObject(item)) {
-			const { json, binary, pairedItem } = item;
-			return {
-				json,
-				...(binary && { binary }),
-				...(pairedItem !== undefined && { pairedItem }),
-			};
-		}
-		return { json: item };
-	}) as INodeExecutionData[];
-}
-
-/**
- * Normalize pin data by ensuring all items have the json key wrapper.
- */
-function normalizePinData(data: IPinData): IPinData {
-	return Object.keys(data).reduce<IPinData>((accu, nodeName) => {
-		accu[nodeName] = data[nodeName].map((item) => {
-			if (!isJsonKeyObject(item)) {
-				return { json: item };
-			}
-			return item;
-		});
-		return accu;
-	}, {});
-}
-
-/**
- * Calculate the byte size of pin data.
- */
-export function getPinDataSize(
-	pinData: Record<string, string | INodeExecutionData[]> = {},
-): number {
-	return Object.values(pinData).reduce<number>((acc, value) => {
-		return acc + stringSizeInBytes(value);
-	}, 0);
 }
 
 /**
@@ -131,45 +71,22 @@ export function useWorkflowDocumentStore(id: WorkflowDocumentId) {
 		 * Pin Data
 		 */
 
-		const pinData = ref<IPinData>({});
-
-		function setPinData(newPinData: IPinData) {
-			onChange({ name: 'setPinData', payload: { pinData: newPinData } });
-		}
-
-		function pinNodeData(nodeName: string, data: INodeExecutionData[]) {
-			onChange({ name: 'pinNodeData', payload: { nodeName, data } });
-		}
-
-		function unpinNodeData(nodeName: string) {
-			onChange({ name: 'unpinNodeData', payload: { nodeName } });
-		}
-
-		function renamePinDataNode(oldName: string, newName: string) {
-			onChange({ name: 'renamePinDataNode', payload: { oldName, newName } });
-		}
-
-		function pinDataByNodeName(nodeName: string): INodeExecutionData[] | undefined {
-			if (!pinData.value[nodeName]) return undefined;
-
-			return pinData.value[nodeName].map((item) => item.json) as INodeExecutionData[];
-		}
-
-		/** Returns a mutable shallow copy of all pin data, bypassing the readonly wrapper. */
-		function getPinDataSnapshot(): IPinData {
-			return { ...pinData.value };
-		}
-
-		/** Returns pin data for a specific node, bypassing the readonly wrapper. */
-		function getNodePinData(nodeName: string): INodeExecutionData[] | undefined {
-			return pinData.value[nodeName];
-		}
+		const {
+			pinData,
+			setPinData,
+			pinNodeData,
+			unpinNodeData,
+			renamePinDataNode,
+			pinDataByNodeName,
+			getPinDataSnapshot,
+			getNodePinData,
+		} = useWorkflowDocumentPinData();
 
 		/**
-		 * Handle actions in a CRDT like manner
+		 * Handle tag actions in a CRDT like manner
 		 */
 
-		function onChange(action: WorkflowDocumentAction) {
+		function onChange(action: TagAction) {
 			if (action.name === 'setTags') {
 				tags.value = action.payload.tags;
 			} else if (action.name === 'addTags') {
@@ -177,45 +94,6 @@ export function useWorkflowDocumentStore(id: WorkflowDocumentId) {
 				tags.value = Array.from(uniqueTags);
 			} else if (action.name === 'removeTag') {
 				tags.value = tags.value.filter((tag) => tag !== action.payload.tagId);
-			} else if (action.name === 'setPinData') {
-				const normalized = normalizePinData(action.payload.pinData);
-				pinData.value = normalized;
-				dataPinningEventBus.emit('pin-data', normalized);
-			} else if (action.name === 'pinNodeData') {
-				const storedData = normalizeNodeExecutionData(action.payload.data);
-				pinData.value = { ...pinData.value, [action.payload.nodeName]: storedData };
-				dataPinningEventBus.emit('pin-data', {
-					[action.payload.nodeName]: storedData,
-				});
-			} else if (action.name === 'unpinNodeData') {
-				const { [action.payload.nodeName]: _, ...rest } = pinData.value;
-				pinData.value = rest;
-				dataPinningEventBus.emit('unpin-data', {
-					nodeNames: [action.payload.nodeName],
-				});
-			} else if (action.name === 'renamePinDataNode') {
-				const { oldName, newName } = action.payload;
-				if (pinData.value[oldName]) {
-					const { [oldName]: renamed, ...rest } = pinData.value;
-					pinData.value = { ...rest, [newName]: renamed };
-				}
-				// Update pairedItem sourceOverwrite references to the old node name
-				Object.values(pinData.value)
-					.flatMap((executionData) =>
-						executionData.flatMap((nodeExecution) =>
-							Array.isArray(nodeExecution.pairedItem)
-								? nodeExecution.pairedItem
-								: [nodeExecution.pairedItem],
-						),
-					)
-					.forEach((pairedItem) => {
-						if (
-							typeof pairedItem === 'number' ||
-							pairedItem?.sourceOverwrite?.previousNode !== oldName
-						)
-							return;
-						pairedItem.sourceOverwrite.previousNode = newName;
-					});
 			}
 		}
 
@@ -226,7 +104,7 @@ export function useWorkflowDocumentStore(id: WorkflowDocumentId) {
 			setTags,
 			addTags,
 			removeTag,
-			pinData: readonly(pinData),
+			pinData,
 			setPinData,
 			pinNodeData,
 			unpinNodeData,
